% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiNode.R
\name{aiNode}
\alias{aiNode}
\title{Taking a tree, the trait data, the number of tips, and trait categories as the input, this function returns the AI of an interesting clade.}
\usage{
aiNode(tree, data, nTip, traits, node, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{nTip}{integer}

\item{traits}{character}

\item{node}{integer}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return the AI index of a given node
}
\description{
Taking a tree, the trait data, the number of tips, and trait categories as the input, this function returns the AI of an interesting clade.
}
