% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiNullModel.R
\name{aiNullModel}
\alias{aiNullModel}
\title{create null distribution using AI statistic}
\usage{
aiNullModel(tree, data, obs, rep, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{obs}{vector}

\item{rep}{numeric}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a matrix
}
\description{
create null distribution using AI statistic
}
