% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiWholeTree.R
\name{aiWholeTree}
\alias{aiWholeTree}
\title{wrapper function to calculate AI index of all internal nodes}
\usage{
aiWholeTree(tree, data, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a named vector of numeric
}
\description{
wrapper function to calculate AI index of all internal nodes
}
