% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calAI.R
\name{calAI}
\alias{calAI}
\title{Given a root node of a subclade, return the proportion of traits for all internal nodes under this subclade}
\usage{
calAI(tree, statsList, nTip, node, aiVector, ...)
}
\arguments{
\item{tree}{phylo}

\item{statsList}{list of double}

\item{nTip}{integer}

\item{node}{integer}

\item{aiVector}{a named vector of numeric}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a named vector of numeric
}
\description{
Given a root node of a subclade, return the proportion of traits for all internal nodes under this subclade
}
