% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyOfTraits.R
\name{frequencyOfTraits}
\alias{frequencyOfTraits}
\title{Traverse the whole tree and return the proportion of traits under each internal nodes}
\usage{
frequencyOfTraits(tree, data, nTip, traits, node, frequencyList, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{nTip}{integer}

\item{traits}{character}

\item{node}{integer}

\item{frequencyList}{list}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a list
}
\description{
Traverse the whole tree and return the proportion of traits under each internal nodes
}
