% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growingP.R
\name{growingP}
\alias{growingP}
\title{test if the result reaches equilibrium}
\usage{
growingP(obs, nullM, ...)
}
\arguments{
\item{obs}{vector}

\item{nullM}{matrix}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a matrix
}
\description{
test if the result reaches equilibrium
}
