% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlightNodes.R
\name{highlightNodes}
\alias{highlightNodes}
\title{highlight nodes}
\usage{
highlightNodes(
  tree,
  data,
  P,
  p_value,
  pointSize,
  highlightSize,
  highlightColor,
  ...
)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{P}{vector}

\item{p_value}{numeric}

\item{pointSize}{numeric}

\item{highlightSize}{numeric}

\item{highlightColor}{character}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a plot
}
\description{
highlight nodes
}
