% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullModel.R
\name{nullModel}
\alias{nullModel}
\title{create null distribution}
\usage{
nullModel(tree, data, obs, rep, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{obs}{vector}

\item{rep}{numeric}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a matrix
}
\description{
create null distribution
}
