% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValue.R
\name{p.value}
\alias{p.value}
\title{This function create the vector of p.value}
\usage{
p.value(res, ...)
}
\arguments{
\item{res}{matrix}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a vector
}
\description{
This function create the vector of p.value
}
