% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValueVector.R
\name{p.valueVector}
\alias{p.valueVector}
\title{create the vector of p.value}
\usage{
p.valueVector(obs, nullM, ...)
}
\arguments{
\item{obs}{vector}

\item{nullM}{matrix}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a vector
}
\description{
create the vector of p.value
}
