% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFrequencyOfTraits.R
\name{plotFrequencyOfTraits}
\alias{plotFrequencyOfTraits}
\title{The function presents the frequency of traits in all internal nodes by pie chart.}
\usage{
plotFrequencyOfTraits(tree, data, pointSize = 2, wt = 0.08, ht = 0.08, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{pointSize}{numeric}

\item{wt}{numeric}

\item{ht}{numeric}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a plot
}
\description{
The function presents the frequency of traits in all internal nodes by pie chart.
}
