% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrowingP.R
\name{plotGrowingP}
\alias{plotGrowingP}
\title{present the result of GrowingP}
\usage{
plotGrowingP(obs, nullM, ...)
}
\arguments{
\item{obs}{vector}

\item{nullM}{matrix}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a plot
}
\description{
present the result of GrowingP
}
