% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNullModel.R
\name{plotNullModel}
\alias{plotNullModel}
\title{The function plots the null distribution.}
\usage{
plotNullModel(res, node = NULL, bw = 0.1, ...)
}
\arguments{
\item{res}{matrix}

\item{node}{numeric}

\item{bw}{numeric}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a plot
}
\description{
The function plots the null distribution.
}
