% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotobsAI.R
\name{plotobsAI}
\alias{plotobsAI}
\title{present the frequency of traits by pie chart}
\usage{
plotobsAI(tree, data, obs, pointSize, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{obs}{vector}

\item{pointSize}{numeric}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a plot
}
\description{
present the frequency of traits by pie chart
}
