% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTreeWithTrait.R
\name{plotTreeWithTrait}
\alias{plotTreeWithTrait}
\title{The function colors the tips by their associated traits.}
\usage{
plotTreeWithTrait(tree, data, pointSize = 2, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{pointSize}{numeric}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a plot
}
\description{
The function colors the tips by their associated traits.
}
