% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psNode.R
\name{psNode}
\alias{psNode}
\title{calculate the parsimony score statistic of a given node}
\usage{
psNode(tree, data, traits, cost, nTip, node, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{traits}{character}

\item{cost}{matrix}

\item{nTip}{integer}

\item{node}{integer}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return the PS vector of a given node
}
\description{
calculate the parsimony score statistic of a given node
}
