% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psNullModel.R
\name{psNullModel}
\alias{psNullModel}
\title{create null distribution using PS statistic}
\usage{
psNullModel(tree, data, obs, cost, rep, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{obs}{vector}

\item{cost}{matrix}

\item{rep}{numeric}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a matrix
}
\description{
create null distribution using PS statistic
}
