% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psWholeTree.R
\name{psWholeTree}
\alias{psWholeTree}
\title{wrapper function to calculate PS of all internal nodes}
\usage{
psWholeTree(tree, data, cost, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{cost}{matrix}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return the PS of all internal nodes as a vector
}
\description{
wrapper function to calculate PS of all internal nodes
}
