% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectMonophyleticGroup.R
\name{selectMonophyleticGroup}
\alias{selectMonophyleticGroup}
\title{get rid of redundant}
\usage{
selectMonophyleticGroup(tree, nodes, ...)
}
\arguments{
\item{tree}{phylo}

\item{nodes}{vector}

\item{...}{further arguments passed to or from other methods}
}
\value{
vector
}
\description{
get rid of redundant
}
