% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleTest.R
\name{simpleTest}
\alias{simpleTest}
\title{Taking a tree and the trait data as the input, this function deploys one test.}
\usage{
simpleTest(tree, data, metric, rep, cost = NULL, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{metric}{"AI" or "PS"}

\item{rep}{numeric}

\item{cost}{matrix}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a matrix
}
\description{
Taking a tree and the trait data as the input, this function deploys one test.
}
