% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSummary.R
\name{testSummary}
\alias{testSummary}
\title{This function summarizes the test result.}
\usage{
testSummary(res, ...)
}
\arguments{
\item{res}{matrix}

\item{...}{further arguments passed to or from other methods}
}
\description{
This function summarizes the test result.
}
