% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipsUnderNode.R
\name{tipsUnderNode}
\alias{tipsUnderNode}
\title{Traverse the whole tree and return the proportion of traits under each internal nodes}
\usage{
tipsUnderNode(tree, node, nTip, tipsVector, ...)
}
\arguments{
\item{tree}{phylo}

\item{node}{integer}

\item{nTip}{integer}

\item{tipsVector}{character}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a vector of character
}
\description{
Traverse the whole tree and return the proportion of traits under each internal nodes
}
