% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverseTest.R
\name{traverseTest}
\alias{traverseTest}
\title{Taking a tree and the trait data as the input, this function deploys multiple tests traversing all the internal nodes.}
\usage{
traverseTest(tree, data, metric, rep, minimumCladeSize = 10, cost = NULL, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{metric}{"AI" or "PS"}

\item{rep}{numeric}

\item{minimumCladeSize}{numeric}

\item{cost}{matrix}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a matrix
}
\description{
Taking a tree and the trait data as the input, this function deploys multiple tests traversing all the internal nodes.
}
