% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiNode.R
\name{aiNode}
\alias{aiNode}
\title{calculate AI index of a given node}
\usage{
aiNode(tree, data, node, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{node}{integer}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return the AI index of a given node
}
\description{
calculate AI index of a given node
}
