% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calAI.R
\name{calAI}
\alias{calAI}
\title{The function returns the AI of each internal node subtended by an interesting clade.}
\usage{
calAI(tree, statsList, nTip, node, aiVector = NULL, ...)
}
\arguments{
\item{tree}{phylo}

\item{statsList}{list of double}

\item{nTip}{integer}

\item{node}{integer}

\item{aiVector}{a named vector of numeric}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a named vector of numeric
}
\description{
The function returns the AI of each internal node subtended by an interesting clade.
}
