% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calAI.R
\name{calAI}
\alias{calAI}
\title{given a root node and the statsList, return the AI index of all internal nodes}
\usage{
calAI(tree, statsList, nTip, node, aiVector = NULL, ...)
}
\arguments{
\item{tree}{phylo}

\item{statsList}{list of double}

\item{nTip}{integer}

\item{node}{integer}

\item{aiVector}{a named vector of numeric}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a named vector of numeric
}
\description{
given a root node and the statsList, return the AI index of all internal nodes
}
