% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyOfTraits.R
\name{frequencyOfTraits}
\alias{frequencyOfTraits}
\title{Taking a tree, the trait data, the number of tips, and trait categories as the input, this function returns the proportion of traits under an interesting clade.}
\usage{
frequencyOfTraits(tree, data, nTip, traits, node, frequencyList = NULL, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{nTip}{integer}

\item{traits}{character}

\item{node}{integer}

\item{frequencyList}{list of integer}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a list
}
\description{
Taking a tree, the trait data, the number of tips, and trait categories as the input, this function returns the proportion of traits under an interesting clade.
}
