% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isParent.R
\name{isParent}
\alias{isParent}
\title{if node1 is parent to node2}
\usage{
isParent(tree, node1, node2)
}
\arguments{
\item{tree}{phylo}

\item{node1}{integer}

\item{node2}{integer}
}
\value{
Boolean
}
\description{
if node1 is parent to node2
}
