% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathToRoot.R
\name{pathToRoot}
\alias{pathToRoot}
\title{This function find the path of a given node to the root.}
\usage{
pathToRoot(tree, node, path = NULL, ...)
}
\arguments{
\item{tree}{phylo}

\item{node}{integer}

\item{path}{vector}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a vector
}
\description{
This function find the path of a given node to the root.
}
