% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFrequencyOfTraits.R
\name{plotFrequencyOfTraits}
\alias{plotFrequencyOfTraits}
\title{present the frequency of traits by pie chart}
\usage{
plotFrequencyOfTraits(tree, data, pointSize, wt, ht, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{pointSize}{numeric}

\item{wt}{numeric}

\item{ht}{numeric}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a plot
}
\description{
present the frequency of traits by pie chart
}
