% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNullModel.R
\name{plotNullModel}
\alias{plotNullModel}
\title{create the vector of p.value}
\usage{
plotNullModel(obs, nullM, P, node, bw, ...)
}
\arguments{
\item{obs}{vector}

\item{nullM}{matrix}

\item{P}{vector}

\item{node}{numeric}

\item{bw}{numeric}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a plot
}
\description{
create the vector of p.value
}
