% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTreeWithTrait.R
\name{plotTreeWithTrait}
\alias{plotTreeWithTrait}
\title{present the input data}
\usage{
plotTreeWithTrait(tree, data, pointSize, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{pointSize}{numeric}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a plot
}
\description{
present the input data
}
