% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psNode.R
\name{psNode}
\alias{psNode}
\title{Taking a tree, the trait data, the number of tips, trait categories and the cost matrix as the input, this function returns the PS vector of an interesting clade.}
\usage{
psNode(tree, data, nTip, traits, cost, node, ...)
}
\arguments{
\item{tree}{phylo}

\item{data}{data.frame}

\item{nTip}{integer}

\item{traits}{character}

\item{cost}{matrix}

\item{node}{integer}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return the PS vector of a given node
}
\description{
Taking a tree, the trait data, the number of tips, trait categories and the cost matrix as the input, this function returns the PS vector of an interesting clade.
}
