% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipsUnderNode.R
\name{tipsUnderNode}
\alias{tipsUnderNode}
\title{Given a node, this function returns the vector of tips under this node.}
\usage{
tipsUnderNode(tree, node, nTip, tipsVector = NULL, ...)
}
\arguments{
\item{tree}{phylo}

\item{node}{integer}

\item{nTip}{integer}

\item{tipsVector}{character}

\item{...}{further arguments passed to or from other methods}
}
\value{
Return a vector of character
}
\description{
Given a node, this function returns the vector of tips under this node.
}
